+++ ui/src/app/modules/config-editor/config-editor.component.ts
@@ -82,6 +82,7 @@ export class ConfigEditorComponent implements OnInit, OnDestroy {
   private isMatterSupported = this.$settings.isFeatureEnabled('matterSupport')
   private isProtocolExternalsOnlyEnabled = this.$settings.isFeatureEnabled('protocolExternalsOnly')
   private isMatterDisableIpv4Enabled = this.$settings.isFeatureEnabled('matterDisableIpv4')
+  private isHapDisableIdentifyingMaterialEnabled = this.$settings.isFeatureEnabled('hapDisableIdentifyingMaterial')
 
   public readonly homebridgeConfig = signal<string>('')
   public readonly originalConfig = signal<string>('')
@@ -648,6 +649,7 @@ export class ConfigEditorComponent implements OnInit, OnDestroy {
       isMatterSupported: this.isMatterSupported,
       isProtocolExternalsOnlyEnabled: this.isProtocolExternalsOnlyEnabled,
       isMatterDisableIpv4Enabled: this.isMatterDisableIpv4Enabled,
+      isHapDisableIdentifyingMaterialEnabled: this.isHapDisableIdentifyingMaterialEnabled,
     });
 
     (window as any).monaco.languages.json.jsonDefaults.setDiagnosticsOptions({
@@ -748,7 +750,7 @@ export class ConfigEditorComponent implements OnInit, OnDestroy {
                       description: this.$translate.instant('status.widget.network.network_interface'),
                     },
                   },
-                  hap: createHapSchema(this.$translate, this.isProtocolExternalsOnlyEnabled, 'main'),
+                  hap: createHapSchema(this.$translate, this.isProtocolExternalsOnlyEnabled, 'main', this.isHapDisableIdentifyingMaterialEnabled),
                   ...this.isMatterSupported
                     ? { matter: createMatterSchema(this.$translate, this.isProtocolExternalsOnlyEnabled, 'main', this.isMatterDisableIpv4Enabled) }
                     : {},